# specs/001-dual-subtitle-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Bilingual Subtitle Studio API
  version: 0.1.0
  description: |
    Dual Subtitle MVP의 REST API 컨트랙트. 모든 응답은 표준 envelope
    `{ success, data?, error?, request_id }` 형식을 따른다.
    상세 사양은 specs/001-dual-subtitle-mvp/plan.md 참조.

servers:
  - url: http://localhost:8000/v1
    description: 로컬 개발 서버

tags:
  - name: jobs
    description: 작업 생성 / 조회 / 취소
  - name: media
    description: 영상 / 자막 다운로드 및 cue 조회
  - name: events
    description: 실시간 상태 push (SSE)

paths:
  /jobs:
    post:
      tags: [jobs]
      summary: 새 작업 생성
      description: |
        YouTube 영상 URL을 받아 작업을 생성한다. 동일 영상 ID의 기존 작업이
        존재하면 신규 생성 없이 해당 작업을 재사용해 200으로 응답한다.
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: 신규 작업 생성됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnvelope"
        "200":
          description: 기존 작업 재사용
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      tags: [jobs]
      summary: 최근 작업 목록
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: cursor
          in: query
          schema: { type: string, nullable: true }
          description: 다음 페이지 cursor (created_at 기반)
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/JobStatus" }
          style: form
          explode: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"

  /jobs/{job_id}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      summary: 작업 상세
      operationId: getJob
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobEnvelope" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [jobs]
      summary: 작업 취소 또는 영구 삭제
      description: |
        상태에 따라 분기 동작한다 — 클라이언트는 분기 처리 불필요.

        - **진행 중** (`pending`, `downloading`, `subtitle_processing`,
          `translating`, `rendering`): cancel 시맨틱 — `failed` / `USER_CANCELLED`
          마킹 + `var/storage/<job_id>/` 산출물 purge. DB row 는 보존.
          spec FR-028 (Clarifications 2026-05-27).
        - **종결** (`completed`, `failed`): hard delete — storage purge + DB row
          영구 제거 (cascade 로 SubtitleTrack / Cue / Asset / JobEvent 등 함께 제거).
          spec FR-030a (Clarifications 2026-05-28).

        응답 envelope `data.action` 으로 어떤 동작이 일어났는지 구분 (`cancelled` |
        `deleted`). 그 외 `data` 키는 VideoJob 필드와 동일.
      operationId: cancelOrDeleteJob
      responses:
        "200":
          description: 취소 또는 삭제 처리됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobActionEnvelope" }
              examples:
                cancelled:
                  summary: 진행 중 작업을 취소
                  value:
                    success: true
                    data:
                      id: 01HX2T0000000000000000000C
                      status: failed
                      error_code: USER_CANCELLED
                      action: cancelled
                    request_id: 01HX2T0000000000000000000D
                deleted:
                  summary: 종결 작업을 영구 삭제
                  value:
                    success: true
                    data:
                      id: 01HX2T0000000000000000000E
                      status: completed
                      action: deleted
                    request_id: 01HX2T0000000000000000000F
        "404":
          $ref: "#/components/responses/NotFound"

  /jobs/{job_id}/subtitles:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [media]
      summary: cue 목록 조회 (원문 / 번역 동시)
      operationId: getSubtitles
      parameters:
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubtitleBundleEnvelope" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 작업이 아직 번역 완료되지 않음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /jobs/{job_id}/download:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [media]
      summary: dual subtitle 파일 다운로드
      operationId: downloadDualSubtitle
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [srt, vtt]
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [source-first, target-first]
            default: source-first
      responses:
        "200":
          description: 다운로드 가능
          content:
            application/x-subrip: {}
            text/vtt: {}
          headers:
            Content-Disposition:
              schema: { type: string }
              description: attachment; filename=...
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 작업이 아직 완료되지 않음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /jobs/{job_id}/video:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [media]
      summary: 영상 스트리밍
      description: HTTP Range 헤더 지원
      operationId: streamVideo
      parameters:
        - name: Range
          in: header
          required: false
          schema: { type: string }
      responses:
        "200":
          description: 전체 파일
          content:
            video/mp4: {}
        "206":
          description: Partial Content
          content:
            video/mp4: {}
        "404":
          $ref: "#/components/responses/NotFound"

  /jobs/{job_id}/events:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [events]
      summary: 실시간 작업 이벤트 (SSE)
      description: |
        Server-Sent Events 스트림. 이벤트 타입 및 payload 스키마는
        contracts/events.md 참고.
      operationId: streamJobEvents
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: string }
      responses:
        "200":
          description: 이벤트 스트림 (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema: { type: string, format: ulid }

  responses:
    BadRequest:
      description: 입력 검증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            invalid_url:
              summary: 유효하지 않은 YouTube URL
              value:
                success: false
                error:
                  code: INVALID_URL
                  message: 유효한 YouTube 영상 URL이 아닙니다.
                request_id: 01HX2T0000000000000000000A
            video_too_long:
              summary: 영상 길이가 120분(7200초)을 초과
              value:
                success: false
                error:
                  code: INVALID_INPUT
                  message: "영상 길이가 120분을 초과합니다 (실제: 8123초). 120분 이하 영상만 처리 가능합니다."
                  details:
                    duration_sec: 8123
                    max_duration_sec: 7200
                request_id: 01HX2T0000000000000000000B
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: 상태 불일치
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RateLimited:
      description: 너무 잦은 요청
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  schemas:
    JobStatus:
      type: string
      enum:
        - pending
        - downloading
        - subtitle_processing
        - translating
        - rendering
        - completed
        - failed

    Language:
      type: string
      enum: [ko, ja]

    CreateJobRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: YouTube 단일 영상 URL
          example: https://www.youtube.com/watch?v=abcdefghijk

    VideoMetadata:
      type: object
      properties:
        title: { type: string, nullable: true }
        channel: { type: string, nullable: true }
        channel_url:
          type: string
          format: uri
          nullable: true
          description: YouTube 채널 페이지 URL — UI 채널명 링크에 사용
          example: https://www.youtube.com/@example
        duration_sec: { type: integer, nullable: true }
        subtitle_source:
          type: string
          enum: [manual, auto]
          nullable: true
      required: [title, channel, channel_url, duration_sec, subtitle_source]

    Job:
      type: object
      required:
        - id
        - source_url
        - youtube_video_id
        - status
        - metadata
        - created_at
        - updated_at
        - reused
      properties:
        id: { type: string, format: ulid }
        source_url: { type: string, format: uri }
        youtube_video_id: { type: string, minLength: 11, maxLength: 11 }
        source_language:
          allOf: [{ $ref: "#/components/schemas/Language" }]
          nullable: true
        target_language:
          allOf: [{ $ref: "#/components/schemas/Language" }]
          nullable: true
        status: { $ref: "#/components/schemas/JobStatus" }
        progress:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
          description: 현재 단계의 0~1 진행률 (단계 전이 시 0으로 리셋)
        error_stage: { type: string, nullable: true }
        error_code: { type: string, nullable: true }
        error_message: { type: string, nullable: true }
        metadata: { $ref: "#/components/schemas/VideoMetadata" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        reused:
          type: boolean
          description: 동일 URL의 기존 작업이 재사용되었는지 여부 (생성 시 응답에만 의미 있음)

    JobEnvelope:
      type: object
      required: [success, request_id]
      properties:
        success: { type: boolean }
        data: { $ref: "#/components/schemas/Job" }
        error: { $ref: "#/components/schemas/ErrorBody" }
        request_id: { type: string }

    JobAction:
      type: string
      enum: [cancelled, deleted]
      description: |
        DELETE /jobs/{job_id} 응답에서 어떤 동작이 일어났는지 표시.
        - cancelled: 진행 중 작업이 USER_CANCELLED 로 마킹됨 (DB row 보존)
        - deleted:   종결 작업이 영구 삭제됨 (DB row + 산출물 제거)

    JobActionEnvelope:
      type: object
      required: [success, request_id]
      description: |
        DELETE /jobs/{job_id} 전용 envelope. data 는 VideoJob 필드 + action 을 포함한다
        (backward compatible — 기존 JobEnvelope 호출자도 status / id 등을 그대로 사용 가능).
      properties:
        success: { type: boolean }
        data:
          allOf:
            - $ref: "#/components/schemas/Job"
            - type: object
              required: [action]
              properties:
                action: { $ref: "#/components/schemas/JobAction" }
        error: { $ref: "#/components/schemas/ErrorBody" }
        request_id: { type: string }

    JobListEnvelope:
      type: object
      required: [success, request_id]
      properties:
        success: { type: boolean }
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Job" }
            next_cursor:
              type: string
              nullable: true
        error: { $ref: "#/components/schemas/ErrorBody" }
        request_id: { type: string }

    SubtitleCue:
      type: object
      required: [sequence, start_ms, end_ms, text]
      properties:
        sequence: { type: integer, minimum: 1 }
        start_ms: { type: integer, minimum: 0 }
        end_ms: { type: integer, minimum: 1 }
        text: { type: string }

    SubtitleBundle:
      type: object
      required:
        - job_id
        - source_language
        - target_language
        - source_cues
        - translated_cues
        - total
        - offset
        - limit
      properties:
        job_id: { type: string, format: ulid }
        source_language: { $ref: "#/components/schemas/Language" }
        target_language: { $ref: "#/components/schemas/Language" }
        source_cues:
          type: array
          items: { $ref: "#/components/schemas/SubtitleCue" }
        translated_cues:
          type: array
          items: { $ref: "#/components/schemas/SubtitleCue" }
        total: { type: integer }
        offset: { type: integer }
        limit: { type: integer }

    SubtitleBundleEnvelope:
      type: object
      required: [success, request_id]
      properties:
        success: { type: boolean }
        data: { $ref: "#/components/schemas/SubtitleBundle" }
        error: { $ref: "#/components/schemas/ErrorBody" }
        request_id: { type: string }

    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계 판독용 코드. 카테고리:
              - INVALID_URL, INVALID_INPUT
              - NOT_FOUND
              - CONFLICT, ILLEGAL_STATE
              - SUBTITLE_NOT_FOUND, SUBTITLE_LANGUAGE_UNSUPPORTED
              - DOWNLOAD_FAILED, TRANSLATION_FAILED, RENDER_FAILED
              - RATE_LIMITED
              - USER_CANCELLED
              - INTERNAL_ERROR
        message:
          type: string
          description: 사용자 친화적 메시지(한국어)
        details:
          type: object
          additionalProperties: true

    ErrorEnvelope:
      type: object
      required: [success, error, request_id]
      properties:
        success:
          type: boolean
          enum: [false]
        error: { $ref: "#/components/schemas/ErrorBody" }
        request_id: { type: string }